+++ src/sources/mod.rs
@@ -1,6 +1,8 @@
 use futures::Future;
 
 pub mod file;
+#[cfg(feature = "systemd")]
+pub mod journald;
 #[cfg(feature = "rdkafka")]
 pub mod kafka;
 pub mod statsd;